- oude saver mixin verwijderd - waarschuwingen opgeruimd - close-all zou nu moeten werken

+++ branches/win-version/Lib/MDevice.cpp
@@ -162,7 +162,7 @@ float MDevice::GetLeading() const
 	return mImpl->GetLeading();
 }
 
-float MDevice::GetLineHeight() const
+int32 MDevice::GetLineHeight() const
 {
 	return mImpl->GetLineHeight();
 }
